Внесены изменения в тесты, добавлена html отчета о тестировании, добавление файлов __init__.py, добавление в комит poetry.lock.
