+++ Tests/test_category.py
@@ -1,17 +1,5 @@
-import pytest
 from src.product import Product
 from src.category import Category
-
-
-def test_product_initialization():
-    """Тестирует корректную инициализацию объекта Product."""
-    product = Product("Ноутбук", "Мощный игровой ноутбук", 99999.99, 10)
-
-    assert product.name == "Ноутбук"
-    assert product.description == "Мощный игровой ноутбук"
-    assert product.price == 99999.99
-    assert product.quantity == 10
-
 
 def test_category_initialization():
     """Тестирует корректную инициализацию объекта Category."""
