+++ src/14.1_main.py
@@ -1,6 +1,7 @@
 from src.category import Category
 from src.product import Product
 
+
 if __name__ == "__main__":
     product1 = Product("Samsung Galaxy S23 Ultra",
                        "256GB, Серый цвет, 200MP камера", 180000.0, 5)
