+++ src/product.py
@@ -1,4 +1,5 @@
 class Product:
+
     def __init__(self, name: str, description: str,
                  price: float, quantity: int):
         self.name = name
